refactoring dal/model code
