+++ dal/model/favorite.gen.go
