+++ dao/comment_test.go
@@ -3,7 +3,7 @@ package dao
 import (
 	"testing"
 
-	"biu-x.org/TikTok/model"
+	"biu-x.org/TikTok/dal/model"
 	"biu-x.org/TikTok/module/config"
 	"biu-x.org/TikTok/module/db"
 	"biu-x.org/TikTok/module/log"
