+++ dao/message.go
@@ -1,8 +1,8 @@
 package dao
 
 import (
+	"biu-x.org/TikTok/dal/model"
 	"biu-x.org/TikTok/dal/query"
-	"biu-x.org/TikTok/model"
 )
 
 // 创建聊天消息
