+++ dao/video.go
@@ -1,11 +1,12 @@
 package dao
 
 import (
+	"time"
+
+	"biu-x.org/TikTok/dal/model"
 	"biu-x.org/TikTok/dal/query"
-	"biu-x.org/TikTok/model"
 	"biu-x.org/TikTok/module/log"
 	"gorm.io/gorm"
-	"time"
 )
 
 // CreateVideo 创建视频记录
