+++ module/middleware/jwt/jwt.go
@@ -8,8 +8,8 @@ import (
 
 	"biu-x.org/TikTok/module/log"
 
+	"biu-x.org/TikTok/dal/model"
 	"biu-x.org/TikTok/dal/query"
-	"biu-x.org/TikTok/model"
 	"github.com/golang-jwt/jwt/v5"
 )
