+++ service/message/action.go
@@ -6,7 +6,7 @@ import (
 	"biu-x.org/TikTok/module/log"
 
 	"biu-x.org/TikTok/dao"
-	"biu-x.org/TikTok/model"
+	"biu-x.org/TikTok/dal/model"
 	"biu-x.org/TikTok/module/response"
 	"github.com/gin-gonic/gin"
 )
