+++ service/user/signup.go
@@ -2,7 +2,7 @@ package user
 
 import (
 	"biu-x.org/TikTok/dal/query"
-	"biu-x.org/TikTok/model"
+	"biu-x.org/TikTok/dal/model"
 	"biu-x.org/TikTok/module/response"
 	"github.com/gin-gonic/gin"
 	"golang.org/x/crypto/bcrypt"
