+++ service/user/user_test.go
@@ -2,7 +2,7 @@ package user
 
 import (
 	"biu-x.org/TikTok/dal/query"
-	"biu-x.org/TikTok/model"
+	"biu-x.org/TikTok/dal/model"
 	"biu-x.org/TikTok/module/config"
 	"biu-x.org/TikTok/module/db"
 	"biu-x.org/TikTok/module/log"
